+++ app/src/main/AndroidManifest.xml
@@ -21,8 +21,9 @@
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
-        <activity android:name=".ui.activity.MainActivity"/>
+        <activity android:name=".ui.activity.LoginActivity"/>
         <activity android:name=".ui.activity.InitiativeActivity"/>
+        <activity android:name=".ui.activity.MainActivity"/>
 
     </application>
