+++ app/src/main/java/com/neiapp/spocan/ui/activity/LoginActivity.java
@@ -1,0 +1,147 @@
+package com.neiapp.spocan.ui.activity;
+
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.Button;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.SignInButton;
+import com.google.android.gms.common.api.ApiException;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.GetTokenResult;
+import com.google.firebase.auth.GoogleAuthProvider;
+import com.neiapp.spocan.R;
+
+
+public class LoginActivity extends AppCompatActivity {
+
+    private static final int RC_SIGN_IN = 1;
+    private SignInButton signInButton;
+    private GoogleSignInClient mGoogleSignInClient;
+    private String TAG = "MainActivity";
+    private FirebaseAuth mAuth;
+    Button mSignOutBtn;
+
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        Intent intent = new Intent(this, MainActivity.class);
+        startActivity(intent);
+
+        setContentView(R.layout.login_user);
+
+        mAuth = FirebaseAuth.getInstance();
+        mSignOutBtn = findViewById(R.id.sign_out);
+        signInButton = findViewById(R.id.sign_in_button);
+
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build();
+
+        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
+
+        signInButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                signIn();
+            }
+        });
+
+        mSignOutBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                signOut();
+            }
+        });
+    }
+
+    private void signIn() {
+        Intent signInIntent = mGoogleSignInClient.getSignInIntent();
+        startActivityForResult(signInIntent, RC_SIGN_IN);
+    }
+
+    private void signOut() {
+        FirebaseAuth.getInstance().signOut();
+        mGoogleSignInClient.signOut();
+        mSignOutBtn.setVisibility(View.INVISIBLE);
+
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        if (requestCode == RC_SIGN_IN) {
+            Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
+            handleSignInResult(task);
+        }
+    }
+
+    private void handleSignInResult(Task<GoogleSignInAccount> completedTask) {
+
+        try {
+            GoogleSignInAccount acc = completedTask.getResult(ApiException.class);
+            Toast.makeText(LoginActivity.this, "Logueado Exitosamente! ", Toast.LENGTH_SHORT).show();
+            firebaseAuthWithGoogle(acc.getIdToken());
+        } catch (ApiException ae) {
+            Toast.makeText(LoginActivity.this, "Hubo un error al loguearse ", Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    private void firebaseAuthWithGoogle(String idToken) {
+        AuthCredential credential = GoogleAuthProvider.getCredential(idToken, null);
+        mAuth.signInWithCredential(credential)
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            Log.d(TAG, "Sign in with credential: Success");
+                            FirebaseUser user = mAuth.getCurrentUser();
+                            //Hacer algo con el usuario si hace falta
+                            updateUI(user);
+                        } else {
+                            Log.w(TAG, "Fallo en el sign in", task.getException());
+                            //Hacer algo para el usuario cuando falla si hace falta
+                            updateUI(null);
+
+                        }
+                    }
+                });
+    }
+
+    private void updateUI(FirebaseUser firebaseUser) {
+        mSignOutBtn.setVisibility(View.VISIBLE);
+        firebaseUser.getIdToken(true).addOnCompleteListener(task ->{
+            GetTokenResult result = task.getResult();
+            String token = result.getToken();
+        });
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        FirebaseUser currentUser = mAuth.getCurrentUser();
+        if (currentUser != null) {
+            updateUI(currentUser);
+        }
+    }
+
+}
