+++ app/src/main/java/com/neiapp/spocan/ui/activity/MainActivity.java
@@ -1,161 +1,29 @@
 package com.neiapp.spocan.ui.activity;
 
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
+import android.os.Bundle;
+
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.navigation.ui.NavigationUI;
 
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-
-import android.widget.LinearLayout;
-
-import android.widget.Toast;
-
-import com.google.android.gms.auth.api.signin.GoogleSignIn;
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
-import com.google.android.gms.auth.api.signin.GoogleSignInClient;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-import com.google.android.gms.common.SignInButton;
-import com.google.android.gms.common.api.ApiException;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-
-
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-
-import com.google.firebase.auth.AuthCredential;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.auth.GetTokenResult;
-import com.google.firebase.auth.GoogleAuthProvider;
 import com.neiapp.spocan.R;
 
 
 public class MainActivity extends AppCompatActivity {
 
-    private static final int RC_SIGN_IN = 1;
-    private SignInButton signInButton;
-    private GoogleSignInClient mGoogleSignInClient;
-    private String TAG = "MainActivity";
-    private FirebaseAuth mAuth;
-    Button mSignOutBtn;
-    LinearLayout mParent;
-
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.login_user);
+        setContentView(R.layout.activity_main);
 
         //botones y navegacion de abajo
         BottomNavigationView bottomNavigationView = findViewById(R.id.bottomNavigationView);
-        NavController navController =  Navigation.findNavController(this, R.id.fragment);
+        NavController navController = Navigation.findNavController(this, R.id.fragment);
         NavigationUI.setupWithNavController(bottomNavigationView, navController);
-
-        mAuth = FirebaseAuth.getInstance();
-        mSignOutBtn = findViewById(R.id.sign_out);
-        signInButton = findViewById(R.id.sign_in_button);
-
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build();
-
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
-
-        signInButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                signIn();
-            }
-        });
-
-        mSignOutBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                signOut();
-            }
-        });
-
-    }
-
-    private void signIn() {
-        Intent signInIntent = mGoogleSignInClient.getSignInIntent();
-        startActivityForResult(signInIntent, RC_SIGN_IN);
-    }
-
-    private void signOut() {
-        FirebaseAuth.getInstance().signOut();
-        mGoogleSignInClient.signOut();
-        mSignOutBtn.setVisibility(View.INVISIBLE);
-
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == RC_SIGN_IN) {
-            Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
-            handleSignInResult(task);
-        }
-    }
-
-    private void handleSignInResult(Task<GoogleSignInAccount> completedTask) {
-
-        try {
-            GoogleSignInAccount acc = completedTask.getResult(ApiException.class);
-            Toast.makeText(MainActivity.this, "Logueado Exitosamente! ", Toast.LENGTH_SHORT).show();
-            firebaseAuthWithGoogle(acc.getIdToken());
-        } catch (ApiException ae) {
-            Toast.makeText(MainActivity.this, "Hubo un error al loguearse ", Toast.LENGTH_SHORT).show();
-        }
-    }
-
-    private void firebaseAuthWithGoogle(String idToken) {
-        AuthCredential credential = GoogleAuthProvider.getCredential(idToken, null);
-        mAuth.signInWithCredential(credential)
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            Log.d(TAG, "Sign in with credential: Success");
-                            FirebaseUser user = mAuth.getCurrentUser();
-                            //Hacer algo con el usuario si hace falta
-                            updateUI(user);
-                        } else {
-                            Log.w(TAG, "Fallo en el sign in", task.getException());
-                            //Hacer algo para el usuario cuando falla si hace falta
-                            updateUI(null);
-
-                        }
-                    }
-                });
-    }
-
-    private void updateUI(FirebaseUser firebaseUser) {
-        mSignOutBtn.setVisibility(View.VISIBLE);
-        firebaseUser.getIdToken(true).addOnCompleteListener(task ->{
-            GetTokenResult result = task.getResult();
-            String token = result.getToken();
-        });
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        FirebaseUser currentUser = mAuth.getCurrentUser();
-        if (currentUser != null) {
-            updateUI(currentUser);
-        }
     }
 
 }
